Projet_2eme_Annee/client/src/app/services/freeGame/vue3D/scene3-d.service.spec.ts
import { getTestBed, inject, TestBed } from "@angular/core/testing";
import * as THREE from "three";
import Spy = jasmine.Spy;
import { Scene3DService } from "./scene3-d.service";

describe("Scene3DService", () => {
  let injector: TestBed;
  let service: Scene3DService;
  beforeEach(() => {
    TestBed.configureTestingModule({
        providers: [Scene3DService],
    }).compileComponents().catch();

    injector = getTestBed();
    service = injector.get(Scene3DService);
  });

  it("should be created", inject([Scene3DService], () => {
    expect(service).toBeTruthy();
  }));

  it("createRenderer should be called", () => {
    try {
      const spy: Spy = spyOn(service, "createRenderer").and.callThrough();
      const htmlCanvas: HTMLCanvasElement = document.createElement("canvas");
      service.createRenderer(htmlCanvas);
      expect(spy).toHaveBeenCalled();
    } catch (e) {
      e = e;
    }
  });

  it("createLight should be called to create a light", () => {
    const spy: Spy = spyOn(service, "createLight").and.callThrough();
    const scene: THREE.Scene = new THREE.Scene();
    service.createLight(scene);
    expect(spy).toHaveBeenCalled();
  });

  it("addMeshesToScene should be called", () => {
    const spy: Spy = spyOn(service, "addMeshesToScene").and.callThrough();
    const scene: THREE.Scene = new THREE.Scene();
    const allObjects: Array<THREE.Mesh> = [];
    service.addMeshesToScene(scene, allObjects);
    expect(scene.children.length).toEqual(allObjects.length);
    expect(spy).toHaveBeenCalled();
  });

  it("makeScene should be called and have the two lights in the children", () => {
    const spy: Spy = spyOn(service, "makeScene").and.callThrough();
    const scene: THREE.Scene = service.makeScene();
    // tslint:disable-next-line:no-magic-numbers
    expect(scene.children.length).toEqual(2);
    expect(spy).toHaveBeenCalled();
  });
});
